// packages/zero/src/zero-schema.gen.ts
const accountTable = {
  name: "account",
  columns: {
    id: {
      type: "string",
      optional: false,
      customType: null as unknown as string,
    },
    userId: {
      type: "string",
      optional: false,
      customType: null as unknown as string,
      serverName: "user_id",
    },
    issuer: {
      type: "string",
      optional: false,
      customType: null as unknown as string,
    },
    accountId: {
      type: "string",
      optional: false,
      customType: null as unknown as string,
      serverName: "account_id",
    },
    providerId: {
      type: "string",
      optional: false,
      customType: null as unknown as string,
      serverName: "provider_id",
    },
    accessToken: {
      type: "string",
      optional: true,
      customType: null as unknown as string,
      serverName: "access_token",
    },
    refreshToken: {
      type: "string",
      optional: true,
      customType: null as unknown as string,
      serverName: "refresh_token",
    },
    accessTokenExpiresAt: {
      type: "number",
      optional: true,
      customType: null as unknown as number,
      serverName: "access_token_expires_at",
    },
    refreshTokenExpiresAt: {
      type: "number",
      optional: true,
      customType: null as unknown as number,
      serverName: "refresh_token_expires_at",
    },
    scope: {
      type: "string",
      optional: true,
      customType: null as unknown as string,
    },
    idToken: {
      type: "string",
      optional: true,
      customType: null as unknown as string,
      serverName: "id_token",
    },
    password: {
      type: "string",
      optional: true,
      customType: null as unknown as string,
    },
    createdAt: {
      type: "number",
      optional: false,
      customType: null as unknown as number,
      serverName: "created_at",
    },
    updatedAt: {
      type: "number",
      optional: false,
      customType: null as unknown as number,
      serverName: "updated_at",
    },
  },
  primaryKey: ["id"],
} as const;
const groupTable = {
  name: "group",
  columns: {
    id: {
      type: "string",
      optional: false,
      customType: null as unknown as string,
    },
    name: {
      type: "string",
      optional: false,
      customType: null as unknown as string,
    },
  },
  primaryKey: ["id"],
} as const;
const groupMemberTable = {
  name: "groupMember",
  columns: {
    groupId: {
      type: "string",
      optional: false,
      customType: null as unknown as string,
      serverName: "group_id",
    },
    userId: {
      type: "string",
      optional: false,
      customType: null as unknown as string,
      serverName: "user_id",
    },
  },
  primaryKey: ["groupId", "userId"],
  serverName: "group_member",
} as const;
const segmentTable = {
  name: "segment",
  columns: {
    id: {
      type: "string",
      optional: false,
      customType: null as unknown as string,
    },
    showId: {
      type: "string",
      optional: false,
      customType: null as unknown as string,
      serverName: "show_id",
    },
    title: {
      type: "string",
      optional: false,
      customType: null as unknown as string,
    },
    type: {
      type: "string",
      optional: false,
      customType: null as unknown as string,
    },
    position: {
      type: "number",
      optional: false,
      customType: null as unknown as number,
    },
    isCurrent: {
      type: "boolean",
      optional: true,
      customType: null as unknown as boolean,
      serverName: "is_current",
    },
  },
  primaryKey: ["id"],
} as const;
const sessionTable = {
  name: "session",
  columns: {
    id: {
      type: "string",
      optional: false,
      customType: null as unknown as string,
    },
    userId: {
      type: "string",
      optional: false,
      customType: null as unknown as string,
      serverName: "user_id",
    },
    token: {
      type: "string",
      optional: false,
      customType: null as unknown as string,
    },
    expiresAt: {
      type: "number",
      optional: false,
      customType: null as unknown as number,
      serverName: "expires_at",
    },
    ipAddress: {
      type: "string",
      optional: true,
      customType: null as unknown as string,
      serverName: "ip_address",
    },
    userAgent: {
      type: "string",
      optional: true,
      customType: null as unknown as string,
      serverName: "user_agent",
    },
    createdAt: {
      type: "number",
      optional: false,
      customType: null as unknown as number,
      serverName: "created_at",
    },
    updatedAt: {
      type: "number",
      optional: false,
      customType: null as unknown as number,
      serverName: "updated_at",
    },
  },
  primaryKey: ["id"],
} as const;
const showTable = {
  name: "show",
  columns: {
    id: {
      type: "string",
      optional: false,
      customType: null as unknown as string,
    },
    title: {
      type: "string",
      optional: false,
      customType: null as unknown as string,
    },
    isActive: {
      type: "boolean",
      optional: true,
      customType: null as unknown as boolean,
      serverName: "is_active",
    },
    features: {
      type: "json",
      optional: true,
      customType: null as unknown as string[],
    },
  },
  primaryKey: ["id"],
} as const;
const showGroupTable = {
  name: "showGroup",
  columns: {
    showId: {
      type: "string",
      optional: false,
      customType: null as unknown as string,
      serverName: "show_id",
    },
    groupId: {
      type: "string",
      optional: false,
      customType: null as unknown as string,
      serverName: "group_id",
    },
  },
  primaryKey: ["showId", "groupId"],
  serverName: "show_group",
} as const;
const userTable = {
  name: "user",
  columns: {
    id: {
      type: "string",
      optional: false,
      customType: null as unknown as string,
    },
    name: {
      type: "string",
      optional: false,
      customType: null as unknown as string,
    },
    isAdmin: {
      type: "boolean",
      optional: true,
      customType: null as unknown as boolean,
      serverName: "is_admin",
    },
    email: {
      type: "string",
      optional: false,
      customType: null as unknown as string,
    },
    emailVerified: {
      type: "boolean",
      optional: false,
      customType: null as unknown as boolean,
      serverName: "email_verified",
    },
    image: {
      type: "string",
      optional: true,
      customType: null as unknown as string,
    },
    username: {
      type: "string",
      optional: true,
      customType: null as unknown as string,
    },
    createdAt: {
      type: "number",
      optional: false,
      customType: null as unknown as number,
      serverName: "created_at",
    },
    updatedAt: {
      type: "number",
      optional: false,
      customType: null as unknown as number,
      serverName: "updated_at",
    },
  },
  primaryKey: ["id"],
} as const;
const verificationTable = {
  name: "verification",
  columns: {
    id: {
      type: "string",
      optional: false,
      customType: null as unknown as string,
    },
    identifier: {
      type: "string",
      optional: false,
      customType: null as unknown as string,
    },
    value: {
      type: "string",
      optional: false,
      customType: null as unknown as string,
    },
    expiresAt: {
      type: "number",
      optional: false,
      customType: null as unknown as number,
      serverName: "expires_at",
    },
    createdAt: {
      type: "number",
      optional: false,
      customType: null as unknown as number,
      serverName: "created_at",
    },
    updatedAt: {
      type: "number",
      optional: false,
      customType: null as unknown as number,
      serverName: "updated_at",
    },
  },
  primaryKey: ["id"],
} as const;
const accountRelationships = {
  user: [
    {
      sourceField: ["userId"],
      destField: ["id"],
      destSchema: "user",
      cardinality: "one",
    },
  ],
} as const;
const sessionRelationships = {
  user: [
    {
      sourceField: ["userId"],
      destField: ["id"],
      destSchema: "user",
      cardinality: "one",
    },
  ],
} as const;
const userRelationships = {
  memberships: [
    {
      sourceField: ["id"],
      destField: ["userId"],
      destSchema: "groupMember",
      cardinality: "many",
    },
  ],
  sessions: [
    {
      sourceField: ["id"],
      destField: ["userId"],
      destSchema: "session",
      cardinality: "many",
    },
  ],
  accounts: [
    {
      sourceField: ["id"],
      destField: ["userId"],
      destSchema: "account",
      cardinality: "many",
    },
  ],
} as const;
const groupRelationships = {
  members: [
    {
      sourceField: ["id"],
      destField: ["groupId"],
      destSchema: "groupMember",
      cardinality: "many",
    },
  ],
  grants: [
    {
      sourceField: ["id"],
      destField: ["groupId"],
      destSchema: "showGroup",
      cardinality: "many",
    },
  ],
} as const;
const groupMemberRelationships = {
  group: [
    {
      sourceField: ["groupId"],
      destField: ["id"],
      destSchema: "group",
      cardinality: "one",
    },
  ],
  user: [
    {
      sourceField: ["userId"],
      destField: ["id"],
      destSchema: "user",
      cardinality: "one",
    },
  ],
} as const;
const segmentRelationships = {
  show: [
    {
      sourceField: ["showId"],
      destField: ["id"],
      destSchema: "show",
      cardinality: "one",
    },
  ],
} as const;
const showRelationships = {
  grants: [
    {
      sourceField: ["id"],
      destField: ["showId"],
      destSchema: "showGroup",
      cardinality: "many",
    },
  ],
  segments: [
    {
      sourceField: ["id"],
      destField: ["showId"],
      destSchema: "segment",
      cardinality: "many",
    },
  ],
} as const;
const showGroupRelationships = {
  show: [
    {
      sourceField: ["showId"],
      destField: ["id"],
      destSchema: "show",
      cardinality: "one",
    },
  ],
  group: [
    {
      sourceField: ["groupId"],
      destField: ["id"],
      destSchema: "group",
      cardinality: "one",
    },
  ],
} as const;
/**
 * The Zero schema object.
 * This type is auto-generated from your Drizzle schema definition.
 */
export const schema = {
  tables: {
    account: accountTable,
    group: groupTable,
    groupMember: groupMemberTable,
    segment: segmentTable,
    session: sessionTable,
    show: showTable,
    showGroup: showGroupTable,
    user: userTable,
    verification: verificationTable,
  },
  relationships: {
    account: accountRelationships,
    session: sessionRelationships,
    user: userRelationships,
    group: groupRelationships,
    groupMember: groupMemberRelationships,
    segment: segmentRelationships,
    show: showRelationships,
    showGroup: showGroupRelationships,
  },
  enableLegacyQueries: false,
  enableLegacyMutators: false,
} as const;
// ...
export const zql = createBuilder(schema);